packages/examples-cli/pyproject.toml
[project]
authors = [
  { name = "A.suzuki", email = "suzu.devworks@gmail.com" },
]
dependencies = [
  "examples-lib",
  "pillow>=10.4.0",
  "pyyaml>=6.0.2",
]
description = "Add your description here"
dynamic = ["version"]
name = "examples-cli"
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
"examples-cli" = "examples_cli:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "types-pillow>=10.2.0.20240822",
  "types-pyyaml>=6.0.12.20240917",
]

[tool.hatch.version]
path = "src/examples_cli/__version__.py"

[tool.uv.sources]
examples-lib = { workspace = true }
